# Customer-Order-Portal/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smart On X Customer Ordering Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login (username or email)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCredentials"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Account disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current portal session
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-username:
    post:
      operationId: forgotUsername
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailInput"
      responses:
        "200":
          description: Always returns success message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsernameOrEmailInput"
      responses:
        "200":
          description: Always returns success message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Wrong current password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── ADMIN AUTH ───────────────────────────────────────────────────
  /admin/auth/login:
    post:
      operationId: adminLogin
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginCredentials"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/auth/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── PRODUCTS (public) ───────────────────────────────────────────
...
  /admin/staff:
    get:
      operationId: listAdminStaff
      tags: [admin]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
    post:
      operationId: createAdminStaff
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/staff/{id}:
    patch:
      operationId: updateAdminStaff
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
  /admin/staff/{id}/reset-password:
    post:
      operationId: resetAdminStaffPassword
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TempPasswordResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        fieldErrors:
          type: object
          additionalProperties:
            type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ─── AUTH SCHEMAS ─────────────────────────────────────────────
    LoginCredentials:
      type: object
      required: [usernameOrEmail, password]
      properties:
        usernameOrEmail:
          type: string
        password:
          type: string
    AdminLoginCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthSession:
      type: object
      required: [userId, roleKey, username, hubspotCompanyId, hubspotContactId, mustChangePassword]
      properties:
        userId:
          type: number
        roleKey:
          type: string
        roleLabel:
          type: string
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        hubspotCompanyId:
          type: string
        hubspotContactId:
          type: string
        mustChangePassword:
          type: boolean
    AdminSession:
      type: object
      required: [adminId, name, email]
      properties:
        adminId:
          type: number
        name:
          type: string
        email:
          type: string
    EmailInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    UsernameOrEmailInput:
      type: object
      required: [usernameOrEmail]
      properties:
        usernameOrEmail:
          type: string
    ResetPasswordInput:
      type: object
      required: [token, newPassword, confirmPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 10
        confirmPassword:
          type: string
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword, confirmPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 10
        confirmPassword:
          type: string
    TempPasswordResponse:
      type: object
      required: [tempPassword]
      properties:
        tempPassword:
          type: string

    # ─── PRODUCT SCHEMAS ──────────────────────────────────────────
...
    AdminUser:
      type: object
      required: [id, name, email, isActive]
      properties:
        id:
          type: number
        name:
          type: string
        email:
          type: string
        isActive:
          type: boolean
        lastLoginAt:
          type: ["string", "null"]
    AdminUserInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          
        password:
          type: string
          minLength: 10
    AdminUserUpdate:
      type: object
      properties:
        name:
          type: string
        isActive:
          type: boolean